Backend - implement signUp.

+++ GetItFreeService/src/main/java/com/getItFree/exception/ResourceConflictException.java
@@ -1,20 +1,22 @@
 package com.getItFree.exception;
 
+import java.math.BigInteger;
+
 public class ResourceConflictException extends RuntimeException {
 
     private static final long serialVersionUID = 1791564636123821405L;
-    private Long resourceId;
+    private BigInteger resourceId;
 
-    public ResourceConflictException(Long resourceId, String message) {
+    public ResourceConflictException(BigInteger resourceId, String message) {
         super(message);
         this.setResourceId(resourceId);
     }
 
-    public Long getResourceId() {
+    public BigInteger getResourceId() {
         return resourceId;
     }
 
-    public void setResourceId(Long resourceId) {
+    public void setResourceId(BigInteger resourceId) {
         this.resourceId = resourceId;
     }
 }
